requirements.txt
fastapi>=0.111.0
pyyaml>=6.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
google-genai>=0.7.0
notion-client>=2.2.1
pydantic>=2.7.0
redis[hiredis]>=5.0
httpx>=0.27.0
python-multipart>=0.0.9
beautifulsoup4>=4.12.0
lxml>=5.2.0
bcrypt>=4.1.0
cryptography>=42.0.0
stripe>=8.0.0
python-docx>=1.1.0
slowapi>=0.1.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
fakeredis>=2.23.0
